requirements.txt
customtkinter
requests
flask
python-dotenv
